refactor(examples): split canonical workbench
